+++ Runtime/Attributes/Dependency/BySelfOrChildAttribute.cs
@@ -1,0 +1,12 @@
+﻿using Fraktal.Framework.DI.Injector.Strategies;
+
+namespace Fraktal.Framework.DI.Injector.Attributes
+{
+    public class BySelfOrChildAttribute : AutoDependencyAttribute
+    {
+        public BySelfOrChildAttribute()
+        {
+            DependencyStrategy = typeof(SelfOrChildrenMatch);
+        }
+    }
+}